Move task functions to modules
